+++ core/src/main/java/com/graphhopper/storage/AbstractDataAccess.java
@@ -44,7 +44,7 @@ public abstract class AbstractDataAccess implements DataAccess
     protected final ByteOrder byteOrder;
     protected final BitUtil bitUtil;
     protected transient boolean closed = false;
-    protected HashMap edgeIdToTFCcount = new HashMap<Integer, HashMap<Integer, Integer>>();
+    protected HashMap edgeIdToTFCcount = new HashMap<Integer, HashMap<Integer, Double>>();
 
     public AbstractDataAccess( String name, String location, ByteOrder order )
     {
@@ -251,7 +251,7 @@ public abstract class AbstractDataAccess implements DataAccess
     }
 
     @Override
-    public void setTrfCnt( int edgeId, int time, int trafficCount )
+    public void setTrfCnt( int edgeId, int time, double trafficCount )
     {
         if (trafficCount > 0)
         {
@@ -260,7 +260,7 @@ public abstract class AbstractDataAccess implements DataAccess
                 ((HashMap) edgeIdToTFCcount.get(edgeId)).put(time, trafficCount);
             } else
             {
-                HashMap<Integer, Integer> hM = new HashMap();
+                HashMap<Integer, Double> hM = new HashMap();
                 hM.put(time, trafficCount);
                 edgeIdToTFCcount.put(edgeId, hM);
             }
@@ -271,18 +271,18 @@ public abstract class AbstractDataAccess implements DataAccess
     }
 
     @Override
-    public int getTrfCnt( int edgeId, int time )
+    public double getTrfCnt( int edgeId, int time )
     {
         if (edgeIdToTFCcount.containsKey(edgeId))
         {
             if (((HashMap) edgeIdToTFCcount.get(edgeId)).containsKey(time))
             {
-                int trfCnt = (int) ((HashMap) edgeIdToTFCcount.get(edgeId)).get(time);
+                double trfCnt = (double) ((HashMap) edgeIdToTFCcount.get(edgeId)).get(time);
                 return trfCnt;
             } else
-                return 0;
+                return 0.0;
 
         } else
-            return 0;
+            return 0.0;
     }
 }
